final code push

+++ bank/src/main/java/enit/rhaddad/api/MakePaymentCommand.java
@@ -3,6 +3,6 @@ package enit.rhaddad.api;
 import java.math.BigDecimal;
 import java.util.UUID;
 
-public record MakePaymentCommand(UUID transactionId,BigDecimal amount,int cardNumber,int cardCode) {
+public record MakePaymentCommand(UUID transactionId,BigDecimal amount,long cardNumber,long cardCode) {
     
 }
